+++ src/Catalog/CompanyListItem.php
@@ -56,6 +56,9 @@ final class CompanyListItem
     /** @var CompanyAddress */
     private $fullAddress;
 
+    /** @var array */
+    private $extra = [];
+
     /**
      * @internal
      *
@@ -79,6 +82,8 @@ final class CompanyListItem
         }
         $this->averageRating = $data['averageRating'];
         $this->fullAddress = new CompanyAddress($data['fullAddress']);
+
+        $this->extra = $data['extra'] ?? [];
     }
 
     /**
@@ -173,5 +178,13 @@ final class CompanyListItem
     public function getAverageRating(): ?int
     {
         return $this->averageRating;
+    }
+
+    /**
+     * @return array
+     */
+    public function getExtra(): array
+    {
+        return $this->extra;
     }
 }